refactor and get this to work like an actual package
