+++ baquet/watchlist.py
@@ -1,13 +1,20 @@
+'''
+From this module, we control data in the watchlist.
+The watchlist houses a list of users and words of interest.
+'''
+
 from pathlib import Path
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy import create_engine
-from sqlalchemy.sql import func
-from datetime import datetime
 
-from baquet.models.watchlist import BASE, WatchlistSQL, WatchwordsSQL
+from .models.watchlist import BASE, WatchlistSQL, WatchwordsSQL
 
 
 class Watchlist:
+    '''
+    From this class, we control data in the watchlist and watchwords.
+    '''
+
     def __init__(self, name):
         self._name = name
         self._conn = self._make_conn()
@@ -24,29 +31,47 @@ class Watchlist:
         return session
 
     def add_watchlist(self, user):
+        '''
+        Add a user to the watchlist.
+        '''
         user = WatchlistSQL(id=user.get_user_id())
         self._conn.merge(user)
         self._conn.commit()
 
+    def remove_watchlist(self, user):
+        '''
+        Remove a user from the watchlist.
+        '''
+        user = self._conn.query(WatchlistSQL).filter(
+            WatchlistSQL.user_id == user.get_user_id()).first()
+        self._conn.delete(user)
+        self._conn.commit()
+
+    def get_watchlist(self):
+        '''
+        Get the watchlist as a list.
+        '''
+        return [user.user_id for user in self._conn.query(WatchlistSQL).all()]
+
     def add_watchword(self, word):
+        '''
+        Add a search term to the watchwords.
+        '''
         word = WatchwordsSQL(word=word)
         self._conn.merge(word)
         self._conn.commit()
 
-    def remove_watchlist(self, user):
-        user = self._conn.query(WatchlistSQL).filter(
-            WatchlistSQL.id == user.get_user_id()).first()
-        self._conn.delete(user)
-        self._conn.commit()
-
     def remove_watchword(self, word):
+        '''
+        Remove a search term from the watchwords.
+        '''
         word = self._conn.query(WatchwordsSQL).filter(
             WatchwordsSQL.word == word).first()
         self._conn.delete(word)
         self._conn.commit()
 
-    def get_watchlist(self):
-        return [user.id for user in self._conn.query(WatchlistSQL).all()]
-
     def get_watchwords(self):
+        '''
+        Get the watchwords as a list.
+        '''
         return [word.word for word in self._conn.query(WatchwordsSQL).all()]
